work in progress commit

+++ src/modprecision.f90
@@ -32,10 +32,11 @@ integer, parameter :: SHR_KIND_R4 = real32  !selected_real_kind( 6) ! 4 byte rea
 integer, parameter :: SHR_KIND_IN = n_i     ! native integer
 
 !! Module specific precisions
-integer, parameter :: field_r = real128 ! Precision for the most common fields u, v, w
+integer, parameter :: field_r = real32  ! Precision for the most common fields u, v, w
                                         ! And all other fields that do not have
                                         ! their own kind and need to interoperate
-integer, parameter :: pois_r  = real128 ! Precision for the poisson solver, p Fp fields
+integer, parameter :: pois_r  = real64  ! Precision for the poisson solver, p Fp fields
                                         ! and all internal fields used in the module
+                                        ! also affects the old fftnew.f90
 
 end module
